complete comment API
