+++ services/commentRedux/commentSlice.ts
@@ -1,7 +1,7 @@
 import {createAsyncThunk} from '@reduxjs/toolkit';
 import axiosInstance from '../axiosInstance';
 import {API} from '../api';
-import {CommentPost} from './commentTypes';
+import {AddCommentDto, CommentPost} from './commentTypes';
 
 export const fetchCommentsByPost = createAsyncThunk<
   CommentPost[],
@@ -15,5 +15,23 @@ export const fetchCommentsByPost = createAsyncThunk<
     return response.data as CommentPost[];
   } catch (err: any) {
     return rejectWithValue(err.response?.data || 'Network error');
+  }
+});
+
+export const fetchAddComment = createAsyncThunk<
+  CommentPost,
+  AddCommentDto,
+  {rejectValue: string}
+>('comments/addComment', async (commentData, {rejectWithValue}) => {
+  try {
+    const response = await axiosInstance.post(
+      API.ADD_COMMENT_POST,
+      commentData,
+    );
+    return response.data;
+  } catch (err: any) {
+    return rejectWithValue(
+      err.response?.data?.message || 'Thêm comment thất bại',
+    );
   }
 });
